+++ path_portable.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package fastwalk
+
+func useForwardSlash() bool {
+	return false
+}
